+++ src/main/java/org/opensearch/knn/index/codec/KNNCodecVersion.java
@@ -13,9 +13,11 @@ import org.apache.lucene.backward_codecs.lucene94.Lucene94Codec;
 import org.apache.lucene.codecs.Codec;
 import org.apache.lucene.backward_codecs.lucene95.Lucene95Codec;
 import org.apache.lucene.backward_codecs.lucene99.Lucene99Codec;
-import org.apache.lucene.codecs.lucene912.Lucene912Codec;
+import org.apache.lucene.backward_codecs.lucene912.Lucene912Codec;
+import org.apache.lucene.codecs.lucene101.Lucene101Codec;
 import org.apache.lucene.codecs.perfield.PerFieldKnnVectorsFormat;
 import org.opensearch.index.mapper.MapperService;
+import org.opensearch.knn.index.codec.KNN10010Codec.KNN10010Codec;
 import org.opensearch.knn.index.codec.KNN80Codec.KNN80CompoundFormat;
 import org.opensearch.knn.index.codec.KNN80Codec.KNN80DocValuesFormat;
 import org.opensearch.knn.index.codec.KNN910Codec.KNN910Codec;
@@ -128,9 +130,23 @@ public enum KNNCodecVersion {
             .knnVectorsFormat(new KNN9120PerFieldKnnVectorsFormat(Optional.ofNullable(mapperService)))
             .build(),
         KNN9120Codec::new
+    ),
+    V_10_01_0(
+        "KNN10010Codec",
+        new Lucene101Codec(),
+        new KNN9120PerFieldKnnVectorsFormat(Optional.empty()),
+        (delegate) -> new KNNFormatFacade(
+            new KNN80DocValuesFormat(delegate.docValuesFormat()),
+            new KNN80CompoundFormat(delegate.compoundFormat())
+        ),
+        (userCodec, mapperService) -> KNN10010Codec.builder()
+            .delegate(userCodec)
+            .knnVectorsFormat(new KNN9120PerFieldKnnVectorsFormat(Optional.ofNullable(mapperService)))
+            .build(),
+        KNN10010Codec::new
     );
 
-    private static final KNNCodecVersion CURRENT = V_9_12_0;
+    private static final KNNCodecVersion CURRENT = V_10_01_0;
 
     private final String codecName;
     private final Codec defaultCodecDelegate;