+++ src/main/java/org/opensearch/knn/index/vectorvalues/KNNVectorValuesFactory.java
@@ -5,11 +5,7 @@
 
 package org.opensearch.knn.index.vectorvalues;
 
-import org.apache.lucene.index.DocValues;
-import org.apache.lucene.index.DocsWithFieldSet;
-import org.apache.lucene.index.FieldInfo;
-import org.apache.lucene.index.LeafReader;
-import org.apache.lucene.index.VectorEncoding;
+import org.apache.lucene.index.*;
 import org.apache.lucene.search.DocIdSetIterator;
 import org.opensearch.knn.common.FieldInfoExtractor;
 import org.opensearch.knn.index.VectorDataType;
@@ -26,9 +22,13 @@ public final class KNNVectorValuesFactory {
      * Returns a {@link KNNVectorValues} for the given {@link DocIdSetIterator} and {@link VectorDataType}
      *
      * @param vectorDataType {@link VectorDataType}
-     * @param docIdSetIterator {@link DocIdSetIterator}
+     * @param knnVectorValues {@link KnnVectorValues}
      * @return {@link KNNVectorValues}
      */
+    public static <T> KNNVectorValues<T> getVectorValues(final VectorDataType vectorDataType, final KnnVectorValues knnVectorValues) {
+        return getVectorValues(vectorDataType, new KNNVectorValuesIterator.DocIdsIteratorValues(knnVectorValues));
+    }
+
     public static <T> KNNVectorValues<T> getVectorValues(final VectorDataType vectorDataType, final DocIdSetIterator docIdSetIterator) {
         return getVectorValues(vectorDataType, new KNNVectorValuesIterator.DocIdsIteratorValues(docIdSetIterator));
     }
@@ -57,19 +57,24 @@ public final class KNNVectorValuesFactory {
      */
     public static <T> KNNVectorValues<T> getVectorValues(final FieldInfo fieldInfo, final LeafReader leafReader) throws IOException {
         final DocIdSetIterator docIdSetIterator;
-        if (fieldInfo.hasVectorValues()) {
-            if (fieldInfo.getVectorEncoding() == VectorEncoding.BYTE) {
-                docIdSetIterator = leafReader.getByteVectorValues(fieldInfo.getName());
-            } else if (fieldInfo.getVectorEncoding() == VectorEncoding.FLOAT32) {
-                docIdSetIterator = leafReader.getFloatVectorValues(fieldInfo.getName());
-            } else {
-                throw new IllegalArgumentException("Invalid Vector encoding provided, hence cannot return VectorValues");
-            }
-        } else {
+        if (!fieldInfo.hasVectorValues()) {
             docIdSetIterator = DocValues.getBinary(leafReader, fieldInfo.getName());
+            final KNNVectorValuesIterator vectorValuesIterator = new KNNVectorValuesIterator.DocIdsIteratorValues(docIdSetIterator);
+            return getVectorValues(FieldInfoExtractor.extractVectorDataType(fieldInfo), vectorValuesIterator);
         }
-        final KNNVectorValuesIterator vectorValuesIterator = new KNNVectorValuesIterator.DocIdsIteratorValues(docIdSetIterator);
-        return getVectorValues(FieldInfoExtractor.extractVectorDataType(fieldInfo), vectorValuesIterator);
+        if (fieldInfo.getVectorEncoding() == VectorEncoding.BYTE) {
+            return getVectorValues(
+                FieldInfoExtractor.extractVectorDataType(fieldInfo),
+                new KNNVectorValuesIterator.DocIdsIteratorValues(leafReader.getByteVectorValues(fieldInfo.getName()))
+            );
+        } else if (fieldInfo.getVectorEncoding() == VectorEncoding.FLOAT32) {
+            return getVectorValues(
+                FieldInfoExtractor.extractVectorDataType(fieldInfo),
+                new KNNVectorValuesIterator.DocIdsIteratorValues(leafReader.getFloatVectorValues(fieldInfo.getName()))
+            );
+        } else {
+            throw new IllegalArgumentException("Invalid Vector encoding provided, hence cannot return VectorValues");
+        }
     }
 
     @SuppressWarnings("unchecked")